remove auto updater menu
