+++ package.json
@@ -30,8 +30,8 @@
     "hyperquest": "^2.1.2",
     "level": "^1.4.0",
     "minimist": "^1.2.0",
+    "ms": "^0.7.2",
     "on-load": "^3.2.0",
-    "once": "^1.4.0",
     "open": "0.0.5",
     "prettier-bytes": "^1.0.3",
     "sheetify": "^6.0.0",
